Arrumando detalhes da TAL
Organizar os arquivos em Src e Inc, e colocar header guards extras para que apenas a parte desejada da TAL seja compilada
